pyproject.toml
[project]
name = "microtelegram"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11.10"
dependencies = [
    "fastapi>=0.115.8",
    "langchain-community>=0.3.21",
    "langchain>=0.3.23",
    "marshmallow-sqlalchemy>=1.4.1",
    "ollama>=0.4.7",
    "python-dotenv>=1.1.0",
    "requests>=2.32.3",
    "sqlalchemy>=2.0.38",
    "telebot>=0.0.5",
    "uvicorn>=0.34.0",
    "langchain-experimental>=0.3.4",
    "langchain-together>=0.3.0",
]
